+++ R/parallel_ui.R
@@ -1,3 +1,52 @@
+#' @title Function \code{batchtools_drake_tmpl_file}
+#' @description Write the batchtools template file
+#' from one of the examples. If there are multiple files,
+#' only the first one (alphabetically) is written.
+#' @export
+#' @seealso \code{\link{drake_examples}}, \code{\link{drake_example}},
+#' \code{\link{shell_file}}
+#' @return \code{NULL} is returned,
+#' but a batchtools template file is written.
+#' @param example Name of the drake example
+#' from which to take the template file.
+#' Must be listed in \code{\link{drake_examples}()}.
+#' @param to Character vector, where to write the file.
+#' @param overwrite Logical, whether to overwrite an existing file of the
+#' same name.
+#' @examples
+#' \dontrun{
+#' load_basic_example() # Load the canonical example.
+#' # List the drake examples. Only some have template files.
+#' drake_examples()
+#' # Write the batchtools template file from the SLURM example.
+#' batchtools_drake_tmpl_file("slurm") # Writes batchtools.slurm.tmpl.
+#' # Find batchtools.slurm.tmpl with the rest of the example's files.
+#' drake_example("slurm") # Writes a new 'slurm' folder with more files.
+#' # Run the basic example with a
+#' # SLURM-powered parallel backend. Requires SLURM.
+#' library(future.batchtools)
+#' future::plan(batchtools_slurm(template = "batchtools.slurm.tmpl"))
+#' make(my_plan, parallelism = "future_lapply")
+#' }
+batchtools_drake_tmpl_file <- function(
+  example = drake::drake_examples(),
+  to = getwd(),
+  overwrite = FALSE
+){
+  example <- match.arg(example)
+  dir <- system.file(file.path("examples", example), package = "drake",
+    mustWork = TRUE)
+  files <- list.files(dir)
+  template_files <- files[grepl("\\.tmpl$", files)]
+  if (!length(template_files)){
+    stop("No template files found for the ", example, " example.")
+  }
+  file <- file.path(dir, template_files[1])
+  file.copy(from = dir, to = to,
+    overwrite = overwrite, recursive = TRUE)
+  invisible()
+}
+
 #' @title Function \code{parallelism_choices}
 #' @description List the types of supported parallel computing.
 #' @export
@@ -296,7 +345,8 @@ max_useful_jobs <- function(
 #' Use this option to run your project in parallel on a computing cluster
 #' or supercomputer.
 #' @seealso \code{\link{make}}, \code{\link{max_useful_jobs}},
-#' \code{\link{parallelism_choices}}
+#' \code{\link{parallelism_choices}}, \code{\link{batchtools_drake_tmpl_file}},
+#' \code{\link{drake_example}}, \code{\link{drake_examples}}
 #' @export
 #' @return The return value of the call to \code{file.copy()} that
 #' wrote the shell file.